refactor(ast): rename definePrinter to createPrinter and inline the factory
Inline the generic `createPrinterFactory` into the renamed `createPrinter`,
the single helper for building a schema printer. The factory was kept
separate so a printer could be keyed by a field other than `node.type`,
but the last such consumer no longer needs it, so the schema-keyed body
is now inlined. The builder shape, options, and `Printer` result are
unchanged.
