+++ packages/ast/src/types.ts
@@ -3,6 +3,6 @@ export type { DistributiveOmit, NodeDef } from './defineNode.ts'
 export type { InferSchemaNode, ParserOptions } from './infer.ts'
 export type * from './nodes/index.ts'
 export type { ParentOf, Visitor, VisitorContext } from './visitor.ts'
-export type { Printer, PrinterFactoryOptions, PrinterPartial } from './definePrinter.ts'
+export type { Printer, PrinterFactoryOptions, PrinterPartial } from './createPrinter.ts'
 export type { OperationParamsResolver } from './utils/operationParams.ts'
 export type { Enforce, Macro } from './defineMacro.ts'
